requirements.txt
python-dotenv
pangram
colorspacious
